refactor: extract shared skill-file logic, fix codex init path (#532)
Extract duplicated skill-file-reading code from ClaudeAdapter and
CodexAgentAdapter into a shared prepend_skill_files() helper in
adapters.rs. Fix guided init codex path to use Codex::builder() for
consistent binary resolution instead of raw Command::new("codex").
